% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver.R
\docType{data}
\name{liver}
\alias{liver}
\title{Liver Cancer Data}
\format{
A data frame with 1,736 observations and 10 variables:
\describe{
\item{ID}{Unique patient identifier}
\item{time}{Survival time in months}
\item{status}{Event indicator: death due to liver cancer = 1, censored = 0}
\item{sex}{Sex of patient: male = 1, female = 0}
\item{age}{Scaled age of diagnosis}
\item{medh}{Scaled median household income of the subject}
\item{race}{Race: white = 1, other = 0}
\item{grade}{Pathological grade of liver cancer (1 = I, 2 = II, 3 = III, 4 = IV)}
\item{chemo}{Chemotherapy: yes = 1, no = 0}
\item{radio}{Radiation: yes = 1, no = 0}
}
}
\usage{
data(liver)
}
\description{
A sample of 1,736 patients who have been diagnosed with liver cancer between 2012 and 2016, whose cancer grades are well identified. Available individual-level covariates include age at diagnosis, sex, race, grade of liver cancer, median household income, radiation indicator, and chemotherapy indicator. The grade of disease is categorized into four levels: well-differentiated (Grade I), moderately differentiated (Grade II), poorly differentiated (Grade III), and undifferentiated/anaplastic (Grade IV).
}
\examples{
data(liver)
head(liver)
}
