% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsMCR.R
\name{residuals.MCR}
\alias{residuals.MCR}
\title{Compute residuals for MCR model}
\usage{
\method{residuals}{MCR}(object, type = c("cox-snell", "quantile"), ...)
}
\arguments{
\item{object}{An object of class \code{MCR}, typically returned from \code{\link{MCRfit}}.}

\item{type}{Type of residual.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals of class \code{"MCRresiduals"}. Printing it shows
the first six residuals and a summary; use \code{as.vector()} to drop the class.
}
\description{
This function computes Global Cox-Snell and randomized quantile residuals for objects of class \code{MCR}.
}
\examples{
data(liver)
liver$sex <- factor(liver$sex)
liver$grade <- factor(liver$grade)
liver$radio <- factor(liver$radio)
liver$chemo <- factor(liver$chemo)

model <- MCRfit(
  survival::Surv(time, status) ~ age + sex + grade + radio + chemo |
    age + medh + grade + radio + chemo,
  dist = "loglogistic", link = "plogit", tau = 0.15,
  data = liver
)
residuals(model, type = "quantile")

}
