% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCR.R
\name{plot.MCR}
\alias{plot.MCR}
\title{Plot multiple MCR model fits against Kaplan-Meier curve}
\usage{
\method{plot}{MCR}(...)
}
\arguments{
\item{...}{One or more fitted MCR objects from \code{MCRfit()}.}
}
\value{
A ggplot object with Kaplan-Meier and survival curves for each model.
}
\description{
Plot multiple MCR model fits against Kaplan-Meier curve
}
